# specs/001-create-receipt/contracts/create-receipt.openapi.yaml
openapi: 3.1.0
info:
  title: Cheapskountant receipt
  version: 1.0.0
  description: |
    Transcribes one Polish fiscal receipt from one to five photos, then stores
    that structured receipt in a separate request.
    The body matches .external-resources/receipt-schema.json.
    Storage adds id. Photos are not stored and are not returned.
servers:
  - url: /
paths:
  /api/transcription:
    post:
      operationId: transcribeReceipt
      summary: Transcribe receipt photos into one structured receipt
      description: Does not store a receipt and does not return an id.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [images]
              properties:
                images:
                  type: array
                  minItems: 1
                  maxItems: 5
                  description: |
                    Ordered pages of one receipt. Repeat the images part.
                    Each part is at most 10 MB.
                    Declared types image/jpeg, image/png, and image/webp are accepted.
                    A part with no media type is treated as image/jpeg.
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: The photos were transcribed. Nothing was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredReceipt'
        '400':
          description: invalid submission. The model was not called.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
              example:
                reason: invalid submission
                explanation: At least one receipt image is required.
        '401':
          description: not authorized. The model was not called.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
              example:
                reason: not authorized
                explanation: A valid API key is required.
        '422':
          description: unreadable or incomplete. Nothing was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
              examples:
                unreadable:
                  value:
                    reason: unreadable
                    explanation: The images could not be read as one fiscal receipt.
                incomplete:
                  value:
                    reason: incomplete
                    explanation: The seller tax identifier is missing.
        '503':
          description: transcription unavailable because the provider failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
              example:
                reason: transcription unavailable
                explanation: The transcription service did not respond.
        '504':
          description: transcription unavailable because the 60 second budget expired.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
              example:
                reason: transcription unavailable
                explanation: Transcription did not finish in time.
  /api/receipt:
    post:
      operationId: createReceipt
      summary: Store one structured receipt
      description: Does not accept photos and does not call the model. A supplied id is ignored.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoredReceipt'
      responses:
        '201':
          description: The receipt was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredReceipt'
        '400':
          description: invalid submission. The receipt body is missing or unreadable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
              example:
                reason: invalid submission
                explanation: A receipt body is required.
        '401':
          description: not authorized. Nothing was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
              example:
                reason: not authorized
                explanation: A valid API key is required.
        '422':
          description: incomplete or not one fiscal receipt. Nothing was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
              examples:
                incomplete:
                  value:
                    reason: incomplete
                    explanation: The seller tax identifier is missing.
                unreadable:
                  value:
                    reason: unreadable
                    explanation: The receipt is not one fiscal receipt.
        '500':
          description: storage failed. Nothing remains stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Failure'
              example:
                reason: storage failed
                explanation: The receipt could not be saved.
  /actuator/health:
    get:
      operationId: health
      summary: Public process health
      security: []
      responses:
        '200':
          description: The process can serve traffic.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Shared secret from the API_KEY environment variable.
  schemas:
    Failure:
      type: object
      additionalProperties: false
      required: [reason, explanation]
      properties:
        reason:
          type: string
          enum:
            - invalid submission
            - not authorized
            - unreadable
            - incomplete
            - transcription unavailable
            - storage failed
        explanation:
          type: string
          minLength: 1
    StoredReceipt:
      type: object
      description: |
        receipt-schema.json with an added id.
        Required groups are documentType, seller, receipt, items, taxSummary,
        totals, and payments. Optional groups are source, fiscalData, and
        unparsedLines. Money values are strings with two fractional digits.
      required: [documentType, seller, receipt, items, taxSummary, totals, payments]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: Assigned by receipt storage and returned after a successful store. Omitted by transcription. Ignored when sent to receipt storage.
        documentType:
          type: string
          enum: [fiscal_receipt]
        source:
          $ref: '#/components/schemas/Source'
        seller:
          $ref: '#/components/schemas/Seller'
        receipt:
          $ref: '#/components/schemas/ReceiptHeader'
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LineItem'
        taxSummary:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/TaxSummaryEntry'
        totals:
          $ref: '#/components/schemas/Totals'
        payments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Payment'
        fiscalData:
          $ref: '#/components/schemas/FiscalData'
        unparsedLines:
          type: array
          items:
            type: string
    Source:
      type: object
      additionalProperties: false
      properties:
        fileName:
          type: string
        rawText:
          type: string
    Seller:
      type: object
      additionalProperties: false
      required: [tradeName, taxId]
      properties:
        tradeName:
          type: string
        legalName:
          type: string
        taxId:
          type: string
          pattern: '^[0-9]{10}$'
        bdoNumber:
          type: string
          pattern: '^[0-9]{9}$'
        businessAddress:
          $ref: '#/components/schemas/Address'
        registeredAddress:
          $ref: '#/components/schemas/Address'
    Address:
      type: object
      additionalProperties: false
      required: [street, postalCode, city, countryCode]
      properties:
        street:
          type: string
        postalCode:
          type: string
          pattern: '^[0-9]{2}-[0-9]{3}$'
        city:
          type: string
        countryCode:
          type: string
          pattern: '^[A-Z]{2}$'
    ReceiptHeader:
      type: object
      additionalProperties: false
      required: [number, issuedAt, currency]
      properties:
        number:
          type: string
        issuedAt:
          type: string
          format: date-time
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        orderNumber:
          type: string
    LineItem:
      type: object
      additionalProperties: false
      required: [description, quantity, unitPrice, total, taxCategory]
      properties:
        description:
          type: string
        itemType:
          type: string
          enum: [product, service, packaging, other]
        quantity:
          type: number
          exclusiveMinimum: 0
        unit:
          type: string
        unitPrice:
          $ref: '#/components/schemas/Money'
        total:
          $ref: '#/components/schemas/Money'
        taxCategory:
          type: string
    TaxSummaryEntry:
      type: object
      additionalProperties: false
      required: [taxCategory, taxRate, taxableSales, taxAmount]
      properties:
        taxCategory:
          type: string
        taxRate:
          type: number
          minimum: 0
          maximum: 100
        taxableSales:
          $ref: '#/components/schemas/Money'
        taxAmount:
          $ref: '#/components/schemas/Money'
    Totals:
      type: object
      additionalProperties: false
      required: [taxAmount, grossAmount, amountDue]
      properties:
        taxAmount:
          $ref: '#/components/schemas/Money'
        grossAmount:
          $ref: '#/components/schemas/Money'
        amountDue:
          $ref: '#/components/schemas/Money'
    Payment:
      type: object
      additionalProperties: false
      required: [method, amount]
      properties:
        method:
          type: string
          enum: [cash, card, transfer, voucher, mobile, other]
        amount:
          $ref: '#/components/schemas/Money'
        transactionId:
          type: string
    FiscalData:
      type: object
      additionalProperties: false
      properties:
        cashRegisterCode:
          type: string
        cashierCode:
          type: string
        fiscalDeviceNumber:
          type: string
        verificationHash:
          type: string
        rawIdentificationLine:
          type: string
    Money:
      type: string
      pattern: '^-?[0-9]+\.[0-9]{2}$'